Struct oauth2::authorization_server::ClientData

pub struct ClientData {
    pub client_id: ClientId,
    pub client_type: ClientType,
    pub redirect_uri: Vec<Url>,
    pub credentials: String,
    pub authn_scheme: Option<String>,
}

Client data is registered with the Authorization Service out of band prior to the OAuth 2.0 protocol commencing.

Fields

client_id

Client Identifier. Required. Must be unique across the Authorization Service and so is typically issued by the Authorization Service. Also used to fetch this entire record.

client_type

Client Type. Required, even if your implementation only ever uses one type.

redirect_uri

Redirect URL(s). Required for public clients, required if the client has multiple, otherwise optional, but SHOULD be supplied anyway.

credentials

Client Credentials, serialized. Required, but the details are out of scope.

authn_scheme

Authentication Scheme, serialized. Only required if multiple authentication schemes are implemented and the server needs to know which one this client is using.