Struct oauth2::authz_server::TokenError

pub struct TokenError {
    pub error: TokenErrorCode,
    pub error_description: Option<String>,
    pub error_uri: Option<String>,
}

Fields

error

error_description

error_uri

Methods

impl TokenError

fn as_json(&self) -> String

Serde serialization serializes all fields, even the ones that are None. But the standard suggests they should be left out entirely if they do not apply. This function skips fields that are None. JSON Serde deserializer will deserialize missing fields as None, so we can use serde for the reverse.