Struct oauth2::client_data::ClientData

pub struct ClientData {
    pub client_id: String,
    pub client_type: ClientType,
    pub redirect_uri: Vec<String>,
    pub credentials: String,
    pub authn_scheme: Option<String>,
}

Client data is registered with the Authorization Service prior to the OAuth 2.0 protocol commencing. This can be done with config files for well-known clients. It can be done online as well prior to the OAuth 2.0 protocol proper.

Fields

client_id

Client Identifier. Required. Must be unique across the Authorization Service and so is typically issued by the Authorization Service. Also used to fetch this entire record.

client_type

Client Type. Required, even if your implementation only ever uses one type.

redirect_uri

Redirect URL(s) as Strings. Required for public clients, required if the client has multiple, otherwise optional but SHOULD be supplied anyway. The first one is used if not supplied in the protocol.

credentials

Client Credentials, serialized. Required, but the details are out of scope.

authn_scheme

Authentication Scheme, serialized. Only required if multiple authentication schemes are implemented and the server needs to know which one this client is using.

Methods

impl ClientData

fn http_basic_authentication_generate(&self) -> Authorization<Basic>

fn http_basic_authentication_deconstruct(basic: Basic) -> Result<(String, String), Utf8Error>