Module oauth2::prelude

Crate prelude that should be wildcard-imported by crate users.

Traits

NewType
New type to wrap a more primitive type in a more typesafe manner.

SecretNewType
New type representing a secret value to wrap a more primitive type in a more typesafe manner.