Trait oauth2::prelude::SecretNewType

pub trait SecretNewType<T>: Debug {
    fn new(val: T) -> Self
    where
        Self: Sized;

    fn secret(&self) -> &T;
}

New type representing a secret value to wrap a more primitive type in a more typesafe manner.

Required methods

fn new(val: T) -> Self where
Self: Sized,

Create a new instance to wrap the given val.

fn secret(&self) -> &T

Get the secret contained within this type.

Security Warning

Leaking this value may compromise the security of the OAuth2 flow.

Loading content...