Crate oauth2

Structs

AuthResponse

The response from the initial step of the auth flow (user authorization).

Cfg

Config data for the oauth2 flow. Some fields may be optional depending on the operation you want to perform.

ClientInfo

Client info JSON struct, as created by the Google Cloud Console.

InstalledClientInfo

Client info JSON sub-struct.

Oauth2

The main oauth object. There are 2 main flows you might want to use it for: auth or refresh.

RefreshResult

The token returned by the refresh flow.

Tokens

The tokens returned from the Token Exchange part of the auth flow.