Struct oauth2::StandardTokenResponse

pub struct StandardTokenResponse<EF, TT>where
    EF: ExtraTokenFields,
    TT: TokenType,{ /* private fields */ }

Standard OAuth2 token response.

This struct includes the fields defined in Section 5.1 of RFC 6749, as well as extensions defined by the EF type parameter.

Implementations

impl<EF, TT> StandardTokenResponse<EF, TT>where EF: ExtraTokenFields, TT: TokenType,

pub fn new(access_token: AccessToken, token_type: TT, extra_fields: EF) -> Self

Instantiate a new OAuth2 token response.

pub fn set_access_token(&mut self, access_token: AccessToken)

Set the access_token field.

pub fn set_token_type(&mut self, token_type: TT)

Set the token_type field.

pub fn set_expires_in(&mut self, expires_in: Option<&Duration>)

Set the expires_in field.

pub fn set_refresh_token(&mut self, refresh_token: Option<RefreshToken>)

Set the refresh_token field.

pub fn set_scopes(&mut self, scopes: Option<Vec<Scope>>)

Set the scopes field.

pub fn extra_fields(&self) -> &EF

Extra fields defined by the client application.

pub fn set_extra_fields(&mut self, extra_fields: EF)

Set the extra fields defined by the client application.